refactor(backend): fully refactor backend to more logical structure and to handle async imports and exports in a more concise and FP way
